Made upated labyrinth generation system with multiple types of corridoors.
