--- Assets/Initial_Prototype/Resources/Scripts/GameManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ed09ddc24ac80294fb47c17a866e1144
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
