+++ Assets/Resources/Materials/RoomPrototype.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8522d16c1817212439d1279745f93017
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
