+++ Assets/Resources/Materials/RoomWallPrototpye.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: e5b05ffd138dec040b850e22ea55f49c
-DefaultImporter:
+guid: 69b162807da535943948c2b70474d447
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
