+++ Assets/Resources/Prefabs/LabyrinthGeneration.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65513f9ffee553f4e941062de04d3e92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
