+++ Assets/Resources/Prefabs/LabyrinthGeneration/CrossRoomHallway.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 380898fb762bd104d856e072fe9273d2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
