+++ Assets/Resources/Prefabs/LabyrinthGeneration/IntersectionHallway.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 36282d360e9a38246a10514a5e2b252a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
