+++ Assets/Resources/Prefabs/LabyrinthGeneration/PlaceholderRoom.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a96ad152006c72e4ea20b52846150e20
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
