+++ Assets/Scenes/PurpleParrotTests/Drew_LabyrinthGeneration.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: f2fcc0f64bfce9c4caa0c271363fe8ae
-PrefabImporter:
+guid: f48970ec3b96e6e46b4630c0eae9d2c0
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
